Don't use a global uiCommands dictionary to pass UICommands around, but rather let all viewers and menu's create their own UICommands. Simplifies code and testing.
Added a 'New note with selected categories...' menu item to the category viewer context menu and renamed the 'New task...' menu item to 'New task with selected categories...' for clarity and consistency.

Removed some of the context menu items (such as 'New task...' that was present in every context menu) that weren't really relevant to the particular context.