Introduce a class for task statuses (task.status.TaskStatus). This has as advantages that we don't have to use the strings for statuses anymore and that we can keep stuff that's status specific in the different TaskStatus instances (like labels and bitmap names for the different status).
Don't have methods and classes tied to a specific task state (e.g. TaskList.nrCompletedTasks or uicommand.ViewerHideCompletedTasks) because this is a form of duplication and makes it hard to add or remove task statuses.

Bug fix: update the taskbar icon tool tip text when the user changes the duesoonhours setting.